+++ wgpu-core/src/instance.rs
@@ -67,9 +67,9 @@ pub struct Instance {
 }
 
 impl Instance {
-    pub fn new(name: &str, backends: Backends) -> Self {
-        fn init<A: HalApi>(_: A, mask: Backends) -> Option<A::Instance> {
-            if mask.contains(A::VARIANT.into()) {
+    pub fn new(name: &str, instance_desc: wgt::InstanceDescriptor) -> Self {
+        fn init<A: HalApi>(_: A, instance_desc: &wgt::InstanceDescriptor) -> Option<A::Instance> {
+            if instance_desc.backends.contains(A::VARIANT.into()) {
                 let mut flags = hal::InstanceFlags::empty();
                 if cfg!(debug_assertions) {
                     flags |= hal::InstanceFlags::VALIDATION;
@@ -78,6 +78,7 @@ impl Instance {
                 let hal_desc = hal::InstanceDescriptor {
                     name: "wgpu",
                     flags,
+                    dx12_shader_compiler: instance_desc.dx12_shader_compiler.clone(),
                 };
                 unsafe { hal::Instance::init(&hal_desc).ok() }
             } else {
@@ -88,15 +89,15 @@ impl Instance {
         Self {
             name: name.to_string(),
             #[cfg(feature = "vulkan")]
-            vulkan: init(hal::api::Vulkan, backends),
+            vulkan: init(hal::api::Vulkan, &instance_desc),
             #[cfg(feature = "metal")]
-            metal: init(hal::api::Metal, backends),
+            metal: init(hal::api::Metal, &instance_desc),
             #[cfg(feature = "dx12")]
-            dx12: init(hal::api::Dx12, backends),
+            dx12: init(hal::api::Dx12, &instance_desc),
             #[cfg(feature = "dx11")]
-            dx11: init(hal::api::Dx11, backends),
+            dx11: init(hal::api::Dx11, &instance_desc),
             #[cfg(feature = "gles")]
-            gl: init(hal::api::Gles, backends),
+            gl: init(hal::api::Gles, &instance_desc),
         }
     }